Add files via upload
Here are four files that I used. The barChart and Figure2 files are both javascript files with code to create a barchart. The css files is for the styles and the BarCharts html file puts everything together. BE AWARE that the large november dataset is not present in these files, but is used in the javascript files. Furthermore, I have added a taxi_zones csv file which is used to retrieve more info about a certain pickup location.